Associate analysis with the analysis container
In order to reference the analysis within the dataModel, each one
of them has now a uuid property.

Let's also note that the analysis driver doesn't maintain anymore
a pool of analysis already instantiated.

Each time an analysis is ran, a new instance of the class is
created using the factory.