voAnalysis - API now support analysis parameters
The parameter can be of four types: Bool, Int, Double and Enum
and can also be grouped together.

If an analysis wants to expose parameters, the function setParameterInformation
has to be overloaded.

The value of each added parameter can also be retrieved within the
execute function using the parameter "id".