Fix ctkFunctionCheckCompilerFlags so that successive invocation works
Internally, the macro CMAKE_CXX_ACCEPTS_FLAG calls TRY_COMPILE. To avoid
the cost of compiling the test each time the project is configured, the variable set by
the macro is added to the cache so that following invocation of the macro with
the same variable name skip the compilation step.

For that same reason, ctkFunctionCheckCompilerFlags function appends a unique suffix to
the HAS_FLAG variable. This suffix is created using a 'clean version' of the flag to test