voDelimitedTextPreviewModel  - Update API to expose table data
Internally, the data are copied in OriginalTable and DataTable, if the
user wants to reset the data after some filter have been applied,
it's possible to call resetDataTable().

Note also that the model listen of the Modified Event associate with
dataTable so that the model can update itself when needed.