Add voNormalizationWidget
This widget is composed of a (1)GroupBox and a (2)StackedLayout

Using the internal method "registerLayoutWidget", a new radio button
will be added to the GroupBox. Each time that button will be checked, the
stacked layout will be updated with the widget associated with the
normalization method being selected.

A signal normalizationMethodSelected(methodName) is also emitted each time
a user click/check a radio button.