Add transposeOption parameter to voUtils::transposeTable
By default, the column names are ignored (voUtils::WithoutHeaders)

If option FirstColumnIntoColumnNames is specified, the first column is
expected to contain string. These string will be come the column name
of the transposed table.

If option ColumnNamesIntoFirstColumn is specified, the column name of
the source table will be transposed into a first column of strings.

If option Headers is specified, the combined effect of FirstColumnIntoColumnNames
and ColumnNamesIntoFirstColumn will occur.