Reload dyanmic views on Analysis update
This will ensure that dynamic views are refreshed with new data
when an Analysis is changed via the Update button. The existing
direct call to voDynView::loadDataObject() is no longer necessary,
as it will be called via signal when the page is reloaded.

Also clean up a line of debugging code in voUtils.