voView - Add setDataObjectInternal method
Before that change, most of the views were not keeping a reference
to their associated voDataObject.

Following this commit, the method "setDataObject()" is not virtual anymore.
Instead that method will keep track of the DataObject and then call
setDataObjectInternal()

It will then be possible to implement view functionality that could process
or query the associated dataobject after the view has been loaded.