make OneZoom visualization less picky about its input data
This commit allows us to launch the OneZoom visualization if the user
has selected a group dataObject that contains a tree.
Previously, OneZoom would only launch if the user had selected a tree.
This meant that the user had to drill down into a group dataObject and
select the tree, otherwise nothing would happen.