Only enable analyses that can be performed on the current input
In the analysis menu, grey out any actions that cannot be performed
on the currently selected input.  This change required me to relocate
the information about the type of input that an analysis expects.

Previously, this information was stored within the analysis objects
themselves.  This led to a possible redundancy, as multiple copies
of an analysis could be created, and they would each have their own
copy of this information.

My solution was to relocate this information about analysis input
types into the singleton voAnalysisDriver.