mask the distinction between vtkTable and vtkExtendedTable
Before this commit, TreeHeatmapView & GeigerModelFitting required a
vtkTable as input, while most other classes expect a vtkExtendedTable.
But from the user's point of view, a table is a table.  As of this
commit, these two classes now take a vtkExtendedTable as input.

In order to accommodate these two classes, vtkExtendedTable has a
new "InputData" ivar.  This is simply the unmodified input to the
vtkExtendedTable.  This approach duplicates data, so it's not an
ideal solution, but it's better than what we had before.