improve analysis input data handling
This commit makes it easier for the user to create new analyses.
If they've currently selected an analysis or input, they can now
generate a new analysis on the input data.

Prior to this commit, the user had to navigate away from the existing
analysis results and manually select the associated input data
before they could perform a new analysis on the data.