Compare Trees analysis
New analysis to compare the edge lengths of two trees that share the
same topology.  This analysis required new code to support enum
parameters whose options are determined at run time (not compile time).

I also had to change the TreeHeatmap View to check for the presence
of an array named "differences".  This is not an ideal solution,
but the alternative would be a significant reworking of the Visomics
framework to allow an Analysis to be able to access its View(s).