ENH: Support multiple selection methods.
Add multiple ways to specify the tip selection:
1) select by providing the tip names
2) select by thresholding the tip attributes

ENH:  Output both the subtable and the subtree.

Add the subtable to the output.

Fix the voIOManager::convertTableToExtendedTable() bug.

Allow creating TreeHeatmapItem using "Associate data
with tree" button for the output tree and output table
(it used to requires InputDataType for the tree and the
table).