record analyses in the order they are performed
We now maintain a list of analyses in the order that they were
performed.  This information is used by the save/load workflow
functionality.

Now that we support dynamic parameters, a parameter to an analysis
can actually be a voDataObject.  For this reason, we now have to
be more careful with the order that we perform analyses when we
load a workflow.  Now that we maintain an order list of analyses,
we can be sure that any parameter/input will exist before we
execute an analysis.