fix handling of analyses list
Previously, if a parent data item was closed, its child analyses
were not removed from our analyses list.  This could cause a bus
error if we later attempted to save the current workflow.

Now, when we close a data item, we recursively check all child
items to see if they are analyses, and if so, they are removed
from our list.