remove tableAndTreeMatch check when creating a TreeHeatmap
This change removes the constraint that a table must have a row for
each leaf node of the tree.  Without this constraint, the vtkTreeHeatmapItem
can be used to display incomplete datasets.  The item also seems to
gracefully handle the case where the table & the tree have nothing to
do with each other.