define VisomicsData_DIR for non-SuperBuild case
Set the CMake variable VisomicsData_DIR when we're not performing a
SuperBuild.  This variable gets configured into voTestConfigure.h,
and it used by the "load sample dataset" feature.  Before this commit,
Visomics would segfault when attempting to load the sample data if
it was not built as part of a SuperBuild.