make "associate..." dialog always on top
This commit makes it so the "associate tree with table" dialog stays
on top of the main window of Visomics even after the focus changes
back to the main window.

Previously, this dialog was somewhat inconvenient to use, because
as you select the table and the tree that you'd like to combine,
the "associate" dialog became hidden behind the main window.