change item names in the data browser
Change the way that input data are named within the data browser.
Instead of reflecting the name of the file that they were loaded
from, such data are now named simply after their type (table, tree,
etc).  This makes it easier for us to change the input within a
saved workflow, so that the same workflow can be executed on
different input data.