enable analyses on tree outputs
Before this commit, the resultTree from fitContinuous was created
as a view.  This meant that no further analyses could be performed
on it.

To fix this, I modified GeigerFitContinuous.xml so that this output
tree is created more similarly to the output table.  I also modified
voRemoteCustomAnalysis.cpp so the output tree is created as a
voOutputDataObject.  After doing this, the voDataModel now
automatically treats the output tree as an output (and not a view).

While working on this, I noticed that voCompareTree was also
creating its output tree as a view.  I also fixed this, so that
further analyses could be performed on this output tree as well.