allow users to select multiple inputs
Change the way the DataBrowser operates so that users can
select more than one item at a time.  This change was made
to better support analyses that require multiple inputs.

As a result, changes were also made to the functions involved
in the running of analyses.  Lots of these functions took
in a single DataModelItem as a parameter.  In these cases,
the parameter was changed to accept a list of such items
instead.