more multi-select work
A few more tweaks to get multi-select working nicely for
analyses.

1) Make it so both InputType and OutputType can unconditionally
be used as the inputs to analyses.

2) No longer enable analyses if a child of the selected item
could be considered valid input for it.  I worry that this will
become confusing as we add more grouped data types.  Instead, it
is now the user's responsibility to highlight the data that they
wish to analyze.

This commit also contains a version number bump.