fix ANOVA for sample table
This commit restores our ability to perform the ANOVA analysis on
our sample liver data.  In order to accomplish this, I added a
new attribute to the input element of our analysis XML description
files.  This attribute is called "metadata".  If it is set to false,
then only the "data" portion of the table will be loaded as input
into the analysis.  By default, the entire table is loaded as input,
metadata and all.