resurrect voCorrelationGraphView
This commit reintroduces the correlation graph view on the
output of the cross correlation analysis.  This view now takes
a table as its input and constructs the necessary graph.
Previously, this view accepted a graph as input, which was
created by the analysis.

This commit also contains changes to the .xml format for analyses
descriptions.  This changes allows for an arbitrary number of views
into a single analysis output.  Previously, an analysis output
could only support one or two views.